добавлена смежная таблица

+++ Controllers/eventController.py
@@ -1,5 +1,3 @@
-import os
-
 from flask import Response
 from application import app
 
@@ -11,9 +9,8 @@ from Services.eventService import EventService
 
 from Exceptions.event_not_found_exception import EventNotFoundException
 from Exceptions.event_duplicate_exception import EventDuplicateException
-from Exceptions.EventIdException import EventIdException
+from Exceptions.IncorrectIdException import IncorrectIdException
 from Exceptions.exceptionDetails import ExceptionDetails
-from Exceptions.date_not_found_exception import DateNotFoundException
 
 
 # для валидации
@@ -49,8 +46,9 @@ class EventControler(Resource):
         """
         return jsonify({'events': _eventService.findAllEvents()})
         
-    @app.route('/es/v1/events/<int:event_id>', methods=['GET'])
+    @app.route('/es/v1/events/<event_id>', methods=['GET'])
     def get_event(event_id):
+        print(event_id)
         """
         Параметры: event_id
         
@@ -59,18 +57,23 @@ class EventControler(Resource):
         
         Обрабтка исключения: при отсутствии введенного id в ответ поступает строка с соответствующим сообщением 
         """
+        
         try:
+            if not event_id.isdigit():
+                raise IncorrectIdException('Id должен быть числом')
+            
             event = _eventService.findEvent(event_id)
             return jsonify(event)
-        except EventIdException as exp:
-            return Response(exp.message, status=404)
         except EventNotFoundException as exp:
             logger.error(f"Произошла ошибка при поиске мероприятия с id: {event_id}. Подробности: {exp}")
             return Response(exp.message, status=404)
+        except IncorrectIdException as exc:
+            logger.error(f"Произошла ошибка при вводе id: {event_id}. Подробности: {exc}")
+            return Response(exc.message, status=404)
         
         
     @staticmethod
-    @app.route('/es/v1/events/<int:id>', methods=['DELETE'])
+    @app.route('/es/v1/events/<id>', methods=['DELETE'])
     def delete_event(id):
         """
         Параметры: id
@@ -81,10 +84,13 @@ class EventControler(Resource):
         Обрабтка исключения: при отсутствии введенного id в ответ поступает строка с соответствующим сообщением
         """
         try:
+            if not id.isdigit():
+                raise IncorrectIdException('Id должен быть числом')
+            
             _eventService.deleteEvent(id)
             return jsonify(id)
-        except EventIdException as exp:
-            return Response(exp.message, status=404)
+        except IncorrectIdException as exc:
+            return Response(exc.message, status=400)
         except EventNotFoundException as exp:
             logger.error(f"Произошла ошибка при поиске мероприятия с id: {id}. Подробности: {exp}")
             return Response(exp.message, status=404)
@@ -109,29 +115,22 @@ class EventControler(Resource):
             event = Event()
             event.eventName = request_data['eventName']
             event.description = request_data['description']
-            event.location = request_data['location']
-            event.DateId = request_data['DateId']
-            event.startTime = request_data['startTime']
-            event.endTime = request_data['endTime']
-            event.program = request_data['program']
             event.invitees = request_data['invitees']
     
             _eventService.addEvent(event)
             return jsonify({'events': _eventService.findAllEvents()})
-        except DateNotFoundException as exc:
-            return Response(exc.message, status=404)
         except ValidationError as error:
             # Форматирование ошибки валидации для включения названий полей
             error_details = [format_validation_error(error) for e in error.context] or [format_validation_error(error)]
-            # logger.error(f'{request_data}\n Произошла ошибка ввода. Подробности: {error_details}')
+            logger.error(f'{request_data}\n Произошла ошибка ввода. Подробности: {error_details}')
             return Response(f'Ошибка ввода\nОшибка в поле {error_details[0]["field"]}\n Подробности: {error_details[0]["message"]}')
         except EventDuplicateException as exp:
-            # logger.error(f"{request_data}\nПроизошла ошибка при добавлении мероприятия. Подробности: {exp}")
+            logger.error(f"{request_data}\nПроизошла ошибка при добавлении мероприятия. Подробности: {exp}")
             return Response(exp.message, status=409)
     
     
     @staticmethod
-    @app.route('/es/v1/events/<int:id>', methods=['PUT'])
+    @app.route('/es/v1/events/<id>', methods=['PUT'])
     def update_event(id):
         """
         Параметры: id
@@ -144,23 +143,21 @@ class EventControler(Resource):
         При обновлении записи, если она является дубликатом, ответом будет являтся строка, говорящая о том что это дубликат, код ошибки - 409
         """
         try:
+            if not id.isdigit():
+                raise IncorrectIdException('Id должен быть числом')
+            
             request_data = request.get_json()
             _eventValidator.validate_event(request_data)
             
             event = Event()
             event.eventName = request_data['eventName']
             event.description = request_data['description']
-            event.location = request_data['location']
-            event.DateId = request_data['DateId']
-            event.startTime = request_data['startTime']
-            event.endTime = request_data['endTime']
-            event.program = request_data['program']
             event.invitees = request_data['invitees']
 
         
             _eventService.updateEvent(id, event)
             return jsonify({'events': _eventService.findAllEvents()})
-        except EventIdException as exp:
+        except IncorrectIdException as exp:
             return Response(exp.message, status=404)
         except ValidationError as error:
             # Форматирование ошибки валидации для включения названий полей
